[feat] Add 'obfuscationMode' support to ResChiper
This commit introduces a new feature, 'obfuscationMode,' to the ResChiper Gradle plugin.
This feature allows users to specify the obfuscation mode for resource obfuscation, providing more flexibility in the obfuscation process.

- Added 'obfuscationMode' property to control the obfuscation mode.
- Updated documentation to explain the usage of 'obfuscationMode.'

This enhancement improves the plugin's functionality and provides users with the ability to customize their resource obfuscation according to their needs.

Signed-off-by: Ayush Bisht <ayushbisht5663@gmail.com>